forecast_api/requirements.txt
# FastAPI and ASGI server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database dependencies (from MCP server)
cloud-sql-python-connector>=1.19.0
pg8000>=1.31.5
sqlalchemy>=2.0.0

# Google Cloud authentication
google-auth>=2.28.0

# Environment management
python-dotenv>=1.0.0

# CORS middleware
python-multipart>=0.0.6

# Testing dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
requests>=2.31.0

google-cloud-aiplatform[agent_engines,adk]==1.90.0